backend/python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMuPDF>=1.24.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
Pillow>=9.0.0
boto3>=1.26.0
